Full notice support, fixes #5
Next thing would be to implement a Flashbag, so that a new
Polyfony\Notice would be automatically stored in it, probably thru
Polyfony\Store\Session and that calling __toString would remove the
notice from the flashbag automatically.
So either you display the notice on the same request, and it is
set/removed in the same request, or you set it, you redirect, and it is
display an removed automatically at the next request.

Test URL is /notice/