Polyfony/Router::reverse() added force_tls option
The 4th parameter of Polyfony/Router::reverse() is now $force_tls. It
is false by default.
If set to true, the generated link will use HTTPS:// even if the
current request isn't using it.
This is bypassed in development environment. So that if your
development server has no valid certificate, this won't interfere with
your workflow.